GameUI methods implementation 2

+++ B21 Ex02 Eithan 204311757 Maor 204709950/B21_Ex02/GameUI.cs
@@ -128,7 +128,7 @@ namespace B21_Ex02
 
         private char[,] drawInitialBoard()
         {
-            byte drowingBoardSize = m_Game.BoardSize * 2 + 2;
+            byte drowingBoardSize = (byte)(m_Game.BoardSize * 2 + 2);
             char[,] drowingBoard = new char[drowingBoardSize, drowingBoardSize];
